+++ codegen/models/__init__.py
@@ -2,20 +2,19 @@ import codegen.models.expr as expr
 import codegen.models.statement as stmt
 from codegen.models.ast import AST
 from codegen.models.expr import PredefinedFn
-from codegen.models.memory import Memory, Var, VarScope
 from codegen.models.program import Program
-from codegen.models.types import AST_ID, KEY, SCOPE
+from codegen.models.types import AST_ID, KEY
+from codegen.models.var import DeferredVar, Var, VarScope
 
 __all__ = [
     "expr",
     "AST",
-    "Memory",
     "Var",
     "VarScope",
+    "DeferredVar",
     "Program",
     "stmt",
     "PredefinedFn",
     "AST_ID",
     "KEY",
-    "SCOPE",
 ]