+++ codegen/models/ast.py
@@ -1,15 +1,13 @@
 from __future__ import annotations
 
 from dataclasses import dataclass, field
-from typing import Any, Callable, Optional
+from typing import TYPE_CHECKING, Any, Callable, Optional
 
 from codegen.models.expr import ExceptionExpr, Expr
-from codegen.models.memory import Memory, Var, VarScope
 from codegen.models.statement import (
     AssignStatement,
     BlockStatement,
     Comment,
-    ContinueStatement,
     DefFuncStatement,
     ElseStatement,
     ExceptionStatement,
@@ -17,18 +15,22 @@ from codegen.models.statement import (
     IfStatement,
     ImportStatement,
     LineBreak,
-    NoStatement,
     PythonStatement,
     ReturnStatement,
     SingleExprStatement,
     Statement,
 )
 from codegen.models.types import AST_ID
+from codegen.models.var import DeferredVar, Var, VarScope
+
+if TYPE_CHECKING:
+    from codegen.models.program import Program
 
 
 @dataclass
 class AST:
     id: AST_ID
+    prog: Program
     stmt: Statement
     children: list[AST] = field(default_factory=list)
     _is_frozen: bool = (
@@ -36,8 +38,8 @@ class AST:
     )
 
     @staticmethod
-    def root():
-        return AST("root", NoStatement())
+    def root(prog: Program):
+        return AST(tuple(), prog, BlockStatement())
 
     def __call__(
         self, *args: Callable[[AST], Any] | Statement, return_self: bool = False
@@ -73,11 +75,20 @@ class AST:
     def linebreak(self):
         self._add_stmt(LineBreak())
 
+    def block(self):
+        return self._add_stmt(BlockStatement())
+
     def comment(self, comment: str):
         self._add_stmt(Comment(comment))
 
-    def func(self, name: str, vars: list[Var]):
-        return self._add_stmt(DefFuncStatement(name, vars))
+    def func(self, name: str, vars: list[DeferredVar]):
+        """Define a function. The input variables are deferred vars as they are created for this function (i.e., must not be predefined prior to this function)"""
+        grandchild_id = self.next_grandchild_id()
+        for var in vars:
+            var.set_var(
+                self.prog.create_var(var.name, var.key, grandchild_id, var.force_name)
+            )
+        return self._add_stmt(DefFuncStatement(name, [var.get_var() for var in vars]))
 
     def expr(self, expr: Expr):
         return self._add_stmt(SingleExprStatement(expr))
@@ -85,17 +96,33 @@ class AST:
     def raise_exception(self, expr: ExceptionExpr):
         return self._add_stmt(ExceptionStatement(expr))
 
-    def assign(self, mem: Memory, var: Var, expr: Expr):
-        scope = self.next_var_scope()
-        self._add_stmt(AssignStatement(var, expr))
-        var.set_scope(mem, scope)
+    def assign(self, var: DeferredVar | Var, expr: Expr):
+        """When we assign the variable, we accept that either you create/declare the variable for the first time or reassign it"""
+        if isinstance(var, DeferredVar):
+            # create the variable for the first time
+            real_var = self.prog.create_var(
+                var.name, var.key, self.next_child_id(), var.force_name
+            )
+            var.set_var(real_var)
+            var = real_var
+        else:
+            # the variable is already created -- and we are just reassigning it
+            # we have to make sure that it's accessible from the current scope
+            assert self.prog.is_within_scope(self.next_child_id(), var.scope)
 
-    def for_loop(self, mem: Memory, item: Var, iter: Expr):
-        assert item.scope is None, "The variable is already assigned to a scope"
-        ast = self._add_stmt(ForLoopStatement(item, iter))
-        scope = ast.next_var_scope()
-        item.set_scope(mem, scope)
-        return ast
+        self._add_stmt(AssignStatement(var, expr))
+
+    def for_loop(self, item: DeferredVar, iter: Expr):
+        """When we construct a for-loop, the item (or var) is always declared. We don't allow reassigning the variable in the for-loop."""
+        assert (
+            isinstance(item, DeferredVar) and item.has_not_been_created()
+        ), "The item must be a new variable"
+        var = self.prog.create_var(
+            item.name, item.key, self.next_grandchild_id(), item.force_name
+        )
+        item.set_var(var)
+
+        return self._add_stmt(ForLoopStatement(var, iter))
 
     def if_(self, condition: Expr):
         return self._add_stmt(IfStatement(condition))
@@ -119,31 +146,15 @@ class AST:
             ast, context, stop = fn(ast, context)
         return ast
 
-    def to_python(self, level: int = -1):
+    def to_python(self, level: int = 0):
         """Convert the AST to python code"""
-        if level == -1:
-            assert self.id == "root"
-            return "".join(
-                [
-                    child.to_python(level + 1)
-                    + ("\n" if ci < len(self.children) - 1 else "")
-                    for ci, child in enumerate(self.children)
-                ]
-            )
-
         if isinstance(self.stmt, BlockStatement):
-            prog = ["\t" * level + stmt.to_python() + "\n" for stmt in self.stmt.stmts]
-        else:
-            prog = ["\t" * level + self.stmt.to_python()]
+            # there is no direct statement and the level does not increase after this statement
+            return "\n".join([child.to_python(level) for child in self.children])
 
-        for child in self.children:
-            prog.append("\n")
-            prog.append(child.to_python(level + 1))
-        return "".join(prog)
-
-    def next_var_scope(self) -> VarScope:
-        """Get a scope for the next variable that will be have if it is assigned to this AST"""
-        return VarScope(self.id, len(self.children))
+        prog = ["\t" * level + self.stmt.to_python()]
+        prog.extend((child.to_python(level + 1) for child in self.children))
+        return "\n".join(prog)
 
     def has_statement_between_ast(self, stmtcls: type[Statement], end_ast_id: AST_ID):
         """Check if there is a statement of the given type the current AST and the end_ast (exclusive)"""
@@ -168,20 +179,38 @@ class AST:
 
     def find_ast(self, id: AST_ID) -> Optional[AST]:
         """Find the AST with the given id"""
-        if self.id == id:
+        if len(self.id) > len(id):
+            # the ast that we are looking for is not in the subtree of this ast
+            return None
+
+        for i in range(len(self.id)):
+            if self.id[i] != id[i]:
+                # the ast that we are looking for is not in the subtree of this ast
+                return None
+
+        if len(self.id) == len(id):
             return self
-        for child in self.children:
-            ast = child.find_ast(id)
-            if ast is not None:
-                return ast
-        return None
+
+        ast = self
+        for i in range(len(self.id), len(id)):
+            ast = ast.children[id[i]]
+        return ast
+
+    def next_child_id(self) -> AST_ID:
+        """Get ID for the next child of this AST"""
+        return self.id + (len(self.children),)
+
+    def next_grandchild_id(self) -> AST_ID:
+        """Get ID for the next grandchild of this AST"""
+        return self.id + (len(self.children), 0)
+
+    def next_var_scope(self) -> VarScope:
+        """Get a scope for the next variable that will be have if it is assigned to this AST"""
+        return VarScope(self.id, len(self.children))
 
     def _add_stmt(self, stmt: Statement):
         if self._is_frozen:
             raise Exception("The AST is frozen and cannot be modified")
-        ast = AST(self._next_child_id(), stmt)
+        ast = AST(self.next_child_id(), self.prog, stmt)
         self.children.append(ast)
         return ast
-
-    def _next_child_id(self):
-        return self.id + "__" + str(len(self.children))